# contract/api-spec.yaml
openapi: 3.0.0
info:
  title: Seedance Wizard API（国内版，无支付）
  version: 0.1.0
  description: |
    ComfyUI Seedance 向导的 Node.js 后端 API。
    负责：DeepSeek 导演 / Qwen VL 看图 / Flux 预览 / Seedance 出片代理 / API Key 管理 / 余额追踪。

    认证方式：所有 /api/* 路由（auth 除外）需在 Header 携带：
      Authorization: Bearer sk-seed-<token>
servers:
  - url: http://localhost:3000
    description: 本地开发
  - url: https://api.seedance.cn
    description: 国内生产环境（待定）

# ─────────────────────────────────────────────
# 通用数据结构
# ─────────────────────────────────────────────
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: sk-seed-<token>
  schemas:
    Error:
      type: object
      properties:
        code:    { type: string, example: "INSUFFICIENT_BALANCE" }
        message: { type: string, example: "余额不足，请充值" }
...
    GenerateTask:
      type: object
      properties:
        task_id:        { type: string }
        status:         { type: string, enum: [queued, processing, completed, failed] }
        progress:       { type: integer, minimum: 0, maximum: 100 }
        video_url:      { type: string, nullable: true }
        estimated_cost_fen: { type: integer }
        actual_cost_fen:    { type: integer, nullable: true }
        created_at:     { type: string, format: date-time }
        completed_at:   { type: string, format: date-time, nullable: true }
security:
  - ApiKeyAuth: []

# ─────────────────────────────────────────────
# 路由定义
# ─────────────────────────────────────────────
paths:
...
  /api/video/generate:
    post:
      tags: [Video]
      summary: |
        触发 Seedance 视频生成。
        调用 muapi.ai Seedance API，返回 task_id 供轮询。
        立即扣费（预扣）；失败时退款。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [session_id, mode]
              properties:
                session_id:
                  type: string
                  format: uuid
                mode:
                  type: string
                  enum: [text_to_video, image_to_video]
                  description: 出片模式
                prompt_override:
                  type: string
                  nullable: true
                  description: 手动覆盖 Prompt（不传则用 session 中的 current_prompt）
                image_b64:
                  type: string
                  nullable: true
                  description: I2V 模式的参考图（base64），T2V 模式不传
                aspect_ratio:
                  type: string
                  enum: ["16:9", "9:16", "1:1", "4:3", "3:4", "21:9"]
                  default: "16:9"
                duration:
                  type: integer
                  minimum: 4
                  maximum: 15
                  default: 5
                  description: 视频时长（秒）
                quality:
                  type: string
                  enum: [basic, high]
                  default: high
      responses:
        "202":
          description: 任务已提交，开始异步生成
          content:
            application/json:
              schema:
                type: object
                properties:
                  task_id:            { type: string }
                  estimated_cost_fen: { type: integer, example: 315 }
                  estimated_seconds:  { type: integer, example: 90, description: "预计等待时间（秒）" }
                  balance_after:      { type: integer, description: "预扣后余额（分）" }
        "402":
          description: 余额不足
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/video/{taskId}/status:
    get:
      tags: [Video]
      summary: 查询视频生成状态（轮询接口，建议每 3 秒调一次）
      parameters:
        - name: taskId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GenerateTask" }
        "404":
          description: task_id 不存在
  /api/video/{taskId}/result:
    get:
      tags: [Video]
      summary: 获取已完成任务的视频下载 URL（仅 status=completed 时有效）
      parameters:
        - name: taskId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  video_url:      { type: string, description: "视频直链，有效期 24 小时" }
                  duration_ms:    { type: integer }
                  actual_cost_fen: { type: integer }
        "400":
          description: 任务尚未完成
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  # ── 费用预估（无需余额扣减）──────────────────────
...